This commit might fix CRC issues

+++ uart.h
@@ -7,4 +7,4 @@ void        UART_AllOut();
 void        UART_Push0();
 void        UART_Push1();
 uint8_t     UART_ReadByte(volatile uint8_t *PORTX, uint8_t PinNum);
-uint16_t    UART_CheckCRC(uint8_t MyArray[]);
+uint8_t    UART_CheckCRC(uint8_t MyArray[]);
